+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ### Unreleased
 
+* Implement a flexible config-based UrlProvider that is both able to generate URLs for
+  the most common cases Warden needs them, and to define the application routing to map 
+  those back to controllers.
 * Provide WardenKohanaDependencyFactory::definitions and ::controllerDefinitions
 * Now officially requires ingenerator/tokenista as the default token-generating service
 * Split all actions to individual controllers - these will now need to be routed as HTTPMethodRoutes for each one,